Changes some ints in trigger generator to uint64_t to protect against running out of bits in long runs

+++ sbndaq-artdaq/Generators/ICARUS/ICARUSTriggerUDP.hh
@@ -92,7 +92,7 @@ namespace sbndaq
     //retry for init msg
     int n_init_retries_;
     int n_init_timeout_ms_;
-    int fEventCounter;
+    uint64_t fEventCounter;
     bool use_wr_time_;
     //expected fragments
     int generated_fragments_per_event_;
